# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Melvor EQ2 Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /character:
    get:
      operationId: getCharacter
      tags: [character]
      summary: Get character state
      responses:
        "200":
          description: Character data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    put:
      operationId: updateCharacter
      tags: [character]
      summary: Update character state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterUpdate"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
...
  /character/settings:
    put:
      operationId: updateCombatSettings
      tags: [character]
      summary: Update combat settings (auto-loop, auto-heal, meditation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CombatSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CombatSettings"
...
components:
  schemas:
...
    Character:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        race:
          type: string
        class:
          type: string
        archetype:
          type: string
          enum: [Fighter, Scout, Mage, Priest]
        level:
          type: integer
        xp:
          type: number
        xpToNextLevel:
          type: number
        aaPoints:
          type: integer
        aaPointsSpent:
          type: integer
        gold:
          type: number
        health:
          type: number
        maxHealth:
          type: number
        power:
          type: number
        maxPower:
          type: number
        baseStats:
          $ref: "#/components/schemas/BaseStats"
        gear:
          $ref: "#/components/schemas/GearSet"
        zone:
          type: string
        activeMount:
          type: string
        alignment:
          type: string
          enum: [Qeynos, Freeport, Neutral]
        totalPlayTime:
          type: number
        killCount:
          type: integer
        deathCount:
          type: integer
        autoLoop:
          type: boolean
        autoHeal:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - race
        - class
        - archetype
        - level
        - xp
        - xpToNextLevel
        - aaPoints
        - aaPointsSpent
        - gold
        - health
        - maxHealth
        - power
        - maxPower
        - baseStats
        - gear
        - zone
        - alignment
        - totalPlayTime
        - killCount
        - deathCount
        - autoLoop
        - autoHeal
        - createdAt
        - updatedAt
    CharacterUpdate:
      type: object
      properties:
        name:
          type: string
        zone:
          type: string
        health:
          type: number
        power:
          type: number
        gold:
          type: number
        xp:
          type: number
        level:
          type: integer
        autoLoop:
          type: boolean
        autoHeal:
          type: boolean
    CombatSettings:
      type: object
      description: >
        Partial settings patch — only fields provided are updated.
        All fields are optional so callers can send a single-field patch
        (e.g. { isMeditating: true }) without supplying unrelated settings.
      properties:
        autoLoop:
          type: boolean
        autoHeal:
          type: boolean
        autoPotions:
          type: boolean
        isMeditating:
          type: boolean
          description: >
            Whether the character is in meditation mode. Server ignores
            attempts to enable this while in active combat.
...
    BaseStats:
      type: object
      properties:
        strength:
          type: number
        agility:
          type: number
        stamina:
          type: number
        intelligence:
          type: number
        wisdom:
          type: number
        charisma:
          type: number
      required: [strength, agility, stamina, intelligence, wisdom, charisma]
...
    GearSet:
      type: object
      properties:
        head:
          $ref: "#/components/schemas/Item"
        neck:
          $ref: "#/components/schemas/Item"
        shoulder:
          $ref: "#/components/schemas/Item"
        chest:
          $ref: "#/components/schemas/Item"
        back:
          $ref: "#/components/schemas/Item"
        wrist:
          $ref: "#/components/schemas/Item"
        hands:
          $ref: "#/components/schemas/Item"
        waist:
          $ref: "#/components/schemas/Item"
        legs:
          $ref: "#/components/schemas/Item"
        feet:
          $ref: "#/components/schemas/Item"
        earLeft:
          $ref: "#/components/schemas/Item"
        earRight:
          $ref: "#/components/schemas/Item"
        ringLeft:
          $ref: "#/components/schemas/Item"
        ringRight:
          $ref: "#/components/schemas/Item"
        charm:
          $ref: "#/components/schemas/Item"
        primary:
          $ref: "#/components/schemas/Item"
        secondary:
          $ref: "#/components/schemas/Item"
        ranged:
          $ref: "#/components/schemas/Item"
    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [weapon, armor, accessory, consumable, material, quest, adornment]
        slot:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, legendary, fabled, mythical]
        level:
          type: integer
        stats:
          $ref: "#/components/schemas/ItemStats"
        sellPrice:
          type: number
        buyPrice:
          type: number
        spriteId:
          type: string
        quantity:
          type: integer
        stackable:
          type: boolean
        setName:
          type: string
        adornmentSlots:
          type: integer
        appliedAdornments:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - type
        - slot
        - rarity
        - level
        - stats
        - sellPrice
        - spriteId
    ItemStats:
      type: object
      properties:
        strength:
          type: number
        agility:
          type: number
        stamina:
          type: number
        intelligence:
          type: number
        wisdom:
          type: number
        charisma:
          type: number
        health:
          type: number
        power:
          type: number
        attackRating:
          type: number
        defenseRating:
          type: number
        mitigation:
          type: number
        haste:
          type: number
        critChance:
          type: number
        weaponDamageMin:
          type: number
        weaponDamageMax:
          type: number
        weaponDelay:
          type: number
        doubleAttackChance:
          type: number
        flurryChance:
          type: number
        spellDamageBonus:
          type: number
        healBonus:
          type: number
...